refactor: teamSession 유틸리티를 lib/team으로 이동
- _utils/teamSession.ts → lib/team/teamSession.ts 이동
- 팀 목록/상세/편집 페이지에서 공통으로 사용할 수 있도록 위치 변경
- 관련 import 경로 업데이트
